// src/app/Models/patientDetail.model.ts
export interface PatientDetail{
    patientId: number;
    firstName: string;
    lastName: string;
    gender: string;
    age: number;
    maritalStatus: string;
    dob: Date;
    religion: string;
    phone: string;
    email: string;
    nationality: string;
    state: string;
    occupation: string;
    address: string;
    relativeName: string;
    relativeRelation: string;
    relativePhone: string;
    relativeEmail: string;
    relativeOccupation: string;
    relativeAddress: string;
}